+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [???]
+
+### Added
+
+- `PowerStep.sweet_plot` accepts arguments `pre_trig` and `trig`, indicating
+  first index of power pulse and power pulse duration. If these arguments are
+  provided, position of the pulse is plotted.
+
 ## [1.9.1] - 2026-03-11
 
 ### Changed